# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Synthora API",
    "description": "Multi-model AI orchestration — answers that surpass any single model",
    "version": "2.8.8"
  },
  "paths": {
...
    "/api/socratic/start": {
      "post": {
        "summary": "Socratic Start",
        "description": "Legacy blocking endpoint — kept for backward compatibility.",
        "operationId": "socratic_start_api_socratic_start_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SocraticStartRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SocraticStartResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/socratic/start/stream": {
      "post": {
        "summary": "Socratic Start Stream",
        "description": "方案C: SSE streaming Socratic start — user sees progress in ~10s.\n\nEvents:\n  socratic_stage     — pipeline stage update (recruiting/searching/thinking/analyzing/guiding)\n  socratic_contributor — one expert responded\n  socratic_divergence — divergence analysis ready\n  socratic_ready     — session ready, guide question generated (terminal event)\n  socratic_error     — pipeline error (terminal event)",
        "operationId": "socratic_start_stream_api_socratic_start_stream_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SocraticStartRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/socratic/respond": {
      "post": {
        "summary": "Socratic Respond",
        "operationId": "socratic_respond_api_socratic_respond_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SocraticRespondRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SocraticRespondResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/socratic/reveal": {
      "post": {
        "summary": "Socratic Reveal",
        "operationId": "socratic_reveal_api_socratic_reveal_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SocraticRevealRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SocraticRevealResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CognitiveSnapshot": {
        "properties": {
          "reasoning_depth": {
            "type": "integer",
            "title": "Reasoning Depth"
          },
          "nuance_recognition": {
            "type": "integer",
            "title": "Nuance Recognition"
          },
          "anchoring_detected": {
            "type": "boolean",
            "title": "Anchoring Detected"
          },
          "confirmation_bias": {
            "type": "boolean",
            "title": "Confirmation Bias"
          },
          "blind_spots": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Blind Spots"
          }
        },
        "type": "object",
        "required": [
          "reasoning_depth",
          "nuance_recognition",
          "anchoring_detected",
          "confirmation_bias",
          "blind_spots"
        ],
        "title": "CognitiveSnapshot"
      },
...
      "DivergencePosition": {
        "properties": {
          "stance": {
            "type": "string",
            "title": "Stance"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          },
          "models": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Models"
          }
        },
        "type": "object",
        "required": [
          "stance",
          "summary",
          "models"
        ],
        "title": "DivergencePosition"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SocraticRespondRequest": {
        "properties": {
          "session_id": {
            "type": "string",
            "maxLength": 50,
            "title": "Session Id"
          },
          "message": {
            "type": "string",
            "maxLength": 5000,
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "message"
        ],
        "title": "SocraticRespondRequest"
      },
      "SocraticRespondResponse": {
        "properties": {
          "guide_message": {
            "type": "string",
            "title": "Guide Message"
          },
          "round": {
            "type": "integer",
            "title": "Round"
          },
          "max_rounds": {
            "type": "integer",
            "title": "Max Rounds"
          },
          "latency_ms": {
            "type": "integer",
            "title": "Latency Ms"
          }
        },
        "type": "object",
        "required": [
          "guide_message",
          "round",
          "max_rounds",
          "latency_ms"
        ],
        "title": "SocraticRespondResponse"
      },
      "SocraticRevealDivergenceMap": {
        "properties": {
          "consensus_points": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Consensus Points"
          },
          "divergence_points": {
            "items": {
              "$ref": "#/components/schemas/SocraticRevealDivergencePoint"
            },
            "type": "array",
            "title": "Divergence Points"
          }
        },
        "type": "object",
        "required": [
          "consensus_points",
          "divergence_points"
        ],
        "title": "SocraticRevealDivergenceMap"
      },
      "SocraticRevealDivergencePoint": {
        "properties": {
          "topic": {
            "type": "string",
            "title": "Topic"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "positions": {
            "items": {
              "$ref": "#/components/schemas/DivergencePosition"
            },
            "type": "array",
            "title": "Positions"
          },
          "consensus_ratio": {
            "type": "number",
            "title": "Consensus Ratio"
          },
          "difficulty": {
            "type": "string",
            "title": "Difficulty"
          }
        },
        "type": "object",
        "required": [
          "topic",
          "description",
          "positions",
          "consensus_ratio",
          "difficulty"
        ],
        "title": "SocraticRevealDivergencePoint"
      },
      "SocraticRevealRequest": {
        "properties": {
          "session_id": {
            "type": "string",
            "maxLength": 50,
            "title": "Session Id"
          }
        },
        "type": "object",
        "required": [
          "session_id"
        ],
        "title": "SocraticRevealRequest"
      },
      "SocraticRevealResponse": {
        "properties": {
          "full_answer": {
            "type": "string",
            "title": "Full Answer"
          },
          "divergence_map": {
            "$ref": "#/components/schemas/SocraticRevealDivergenceMap"
          },
          "cognitive_snapshot": {
            "$ref": "#/components/schemas/CognitiveSnapshot"
          },
          "guide_rounds_used": {
            "type": "integer",
            "title": "Guide Rounds Used"
          }
        },
        "type": "object",
        "required": [
          "full_answer",
          "divergence_map",
          "cognitive_snapshot",
          "guide_rounds_used"
        ],
        "title": "SocraticRevealResponse"
      },
      "SocraticStartDivergenceMap": {
        "properties": {
          "consensus_points": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Consensus Points"
          },
          "divergence_count": {
            "type": "integer",
            "title": "Divergence Count"
          },
          "overall_consensus": {
            "type": "number",
            "title": "Overall Consensus"
          }
        },
        "type": "object",
        "required": [
          "consensus_points",
          "divergence_count",
          "overall_consensus"
        ],
        "title": "SocraticStartDivergenceMap"
      },
      "SocraticStartRequest": {
        "properties": {
          "question": {
            "type": "string",
            "maxLength": 10000,
            "title": "Question"
          },
          "locale": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Locale"
          }
        },
        "type": "object",
        "required": [
          "question"
        ],
        "title": "SocraticStartRequest"
      },
      "SocraticStartResponse": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "phase1_latency_ms": {
            "type": "integer",
            "title": "Phase1 Latency Ms"
          },
          "max_guide_rounds": {
            "type": "integer",
            "title": "Max Guide Rounds"
          },
          "divergence_map": {
            "$ref": "#/components/schemas/SocraticStartDivergenceMap"
          },
          "initial_guide": {
            "type": "string",
            "title": "Initial Guide"
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "phase1_latency_ms",
          "max_guide_rounds",
          "divergence_map",
          "initial_guide"
        ],
        "title": "SocraticStartResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}